requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.3.1

# Task Queue & Message Broker
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# YouTube API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Trending Topics & Web Scraping
pytrends==4.9.2
requests==2.31.0

# Text-to-Speech
coqui-tts==0.21.4
gTTS==2.4.0

# Video Processing
moviepy==1.0.3
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Stock Videos API
pexels-api==1.0.2

# Database & ORM
psycopg2-binary==2.9.9

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Error Handling
tenacity==8.2.3

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3
pytest==7.4.3
pytest-django==4.7.0
pytest-celery==0.0.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
